Schedule is ready

+++ Project1/Models/Train.cs
@@ -3,7 +3,7 @@
     public class Train
     {
         public int Id { get; set; }
-        public decimal Capacity { get; set; }
+        public int Capacity { get; set; }
         public int TrainNumber { get; set; }
         public ICollection<Router> Router { get; set; }
         public ICollection<DriverTrainCategory> DriverTrainCategory { get; set; }
